Start to program with TDD

+++ src/main/java/es/fbenavente/documentrelevance/domain/DocumentRelevance.java
@@ -5,13 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class DocumentRelevance {
     private String name;
-    private BigDecimal ranking;
+    private double ranking;
 }
